Using this project.
To use this project you must attach UsingADO.mdf database to a local instance of Sql Server. Once that database is attached, you will need to adjust the connection string in the project to point to the Sql Server instance that you attached the database to. 
Good luck and enjoy the project.